requirements.txt
kafka-python==2.0.2
numpy==1.21.1
opencv-contrib-python==4.5.3.56
pretrainedmodels==0.7.4
PyYAML==5.4.1
segmentation-models-pytorch==0.2.0
six==1.16.0
timm==0.4.12
torch==1.9.0
torchvision==0.10.0